refactor(skills): extract skill IPC handlers into dedicated module
Move 15 skill IPC handlers from main.ts into
src/main/ipcHandlers/skills/, following the established plugin
handlers pattern. Extract updatePluginSkillIdsFromReport helper
and OpenClawSkillReport type into src/main/skills/.
