+++ src/main/ipcHandlers/skills/index.ts
@@ -1,0 +1,2 @@
+export type { SkillHandlerDeps } from './handlers';
+export { registerSkillHandlers } from './handlers';
