+++ src/main/skills/index.ts
@@ -1,0 +1,2 @@
+export { updatePluginSkillIdsFromReport } from './openClawSync';
+export type { OpenClawSkillReport } from './openClawSync';
